+++ comic_dl/honcho.py
@@ -13,6 +13,7 @@ from future import standard_library
 standard_library.install_aliases()
 
 #import urllib as urllib2
+import logging
 from sites.yomanga import yomanga_Url_Check
 from sites.gomanga import gomanga_Url_Check
 from sites.mangafox import mangafox_Url_Check
@@ -27,33 +28,38 @@ from urllib.parse import urlparse
 
 
 
-def url_checker(input_url, current_directory, User_Name, User_Password):
+def url_checker(input_url, current_directory, User_Name, User_Password, logger):
+
+    if logger == "True":
+        logging.basicConfig(format='%(levelname)s: %(message)s', filename="Error Log.log", level=logging.DEBUG)
+        logging.debug("Your URL : %s" % input_url)
 
     domain = urlparse(input_url).netloc
+    logging.debug("Domain : %s" % domain)
 
     if domain in ['mangafox.me']:
-        mangafox_Url_Check(input_url, current_directory)
+        mangafox_Url_Check(input_url, current_directory, logger)
 
     elif domain in ['yomanga.co']:
-        yomanga_Url_Check(input_url, current_directory)
+        yomanga_Url_Check(input_url, current_directory, logger)
 
     elif domain in ['gomanga.co']:
-        gomanga_Url_Check(input_url, current_directory)
+        gomanga_Url_Check(input_url, current_directory, logger)
 
     elif domain in ['bato.to']:
-        batoto_Url_Check(input_url, current_directory, User_Name, User_Password)
+        batoto_Url_Check(input_url, current_directory, User_Name, User_Password, logger)
 
     elif domain in ['kissmanga.com']:
-        kissmanga_Url_Check(input_url, current_directory)
+        kissmanga_Url_Check(input_url, current_directory, logger)
 
     elif domain in ['comic.naver.com']:
-        comic_naver_Url_Check(input_url, current_directory)
+        comic_naver_Url_Check(input_url, current_directory, logger)
 
     elif domain in ['readcomiconline.to']:
-        readcomic_Url_Check(input_url, current_directory)
+        readcomic_Url_Check(input_url, current_directory, logger)
 
     elif domain in ['kisscomic.us']:
-        kissmcomicus_Url_Check(input_url, current_directory)
+        kissmcomicus_Url_Check(input_url, current_directory, logger)
         
     elif domain in ['']:
         print('You need to specify at least 1 URL. Please run : comic-dl -h')