+++ comic_dl/sites/comic_naver.py
@@ -11,8 +11,9 @@ import requests
 from downloader.cookies_required import with_referer as FileDownloader
 from six.moves import range
 from six.moves import input
+import logging
 
-def single_chapter(url,current_directory):
+def single_chapter(url,current_directory, logger):
     
     s = requests.Session()
     headers = {'user-agent': 'Mozilla/5.0 (Windows NT 6.1; WOW64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/54.0.2840.99 Safari/537.36'}
@@ -24,18 +25,21 @@ def single_chapter(url,current_directory):
         #Korean_Name = re.search(r'<h2>(.*?)<span class="wrt_nm">',str(page_source)).group(1)
         Series_Name = re.search(r'titleId=(\d+)',url).group(1)
     except Exception as e:
+        logging.debug("Error in Series Name : %s" % e)
         Series_Name = "Unknown"
 
     try:
         #chapter_number = int(re.search(r'\<span\ class\=\"total\"\>(.\d+)\<\/span\>',page_source_1).group(1))
         chapter_number = re.search(r'&no=(\d+)',url).group(1)
     except Exception as e:
-        print(e)
+        # print(e)
+        logging.debug("Error in Chapter Number : %s" % e)
         chapter_number = 0
     
     img_regex = r'http://imgcomic.naver.net/webtoon/\d+/\d+/.+?\.(?:jpg|png|gif|bmp|JPG|PNG|GIF|BMP)'
 
     img_links = list(re.findall(img_regex,page_source_1))
+    logging.debug("Image Links : %s" % img_links)
     
     Raw_File_Directory = str(Series_Name) +'/'+"Chapter "+str(chapter_number)
 
@@ -51,7 +55,7 @@ def single_chapter(url,current_directory):
     for x,items in enumerate(img_links):
         if not os.path.exists(File_Directory):
             os.makedirs(File_Directory)
-        FileDownloader(str(x+1)+str(items[-4:]),Directory_path,cookies,items,url)
+        FileDownloader(str(x+1)+str(items[-4:]),Directory_path,cookies,items,url, logger)
 
     print('\n')
     print("Completed downloading ",Series_Name)
@@ -59,7 +63,7 @@ def single_chapter(url,current_directory):
 
 
 
-def whole_series(url, current_directory):
+def whole_series(url, current_directory, logger):
     
     
     
@@ -81,11 +85,14 @@ def whole_series(url, current_directory):
     
     for x in range(1,int(first_link)):
         Chapter_Url = "http://comic.naver.com/webtoon/detail.nhn?titleId=%s&no=%s" %(titleId,x)
-        single_chapter(Chapter_Url,current_directory)
+        logging.debug("Chapter URL : %s" % Chapter_Url)
+        single_chapter(Chapter_Url,current_directory, logger)
 
 
 
-def comic_naver_Url_Check(input_url, current_directory):
+def comic_naver_Url_Check(input_url, current_directory, logger):
+    if logger == "True":
+        logging.basicConfig(format='%(levelname)s: %(message)s', filename="Error Log.log", level=logging.DEBUG)
 
     comic_naver_single_regex = re.compile(
         'https?://(?P<host>comic.naver.com)/webtoon/(?P<detail>detail.nhn)\?titleId\=(?P<extra_characters>[\d]+)?(\/|.)')
@@ -99,7 +106,7 @@ def comic_naver_Url_Check(input_url, current_directory):
             match = found.groupdict()
             if match['detail']:
                 url = str(input_url)
-                single_chapter(url, current_directory)
+                single_chapter(url, current_directory, logger)
                 
             else:
                 pass
@@ -109,6 +116,6 @@ def comic_naver_Url_Check(input_url, current_directory):
             match = found.groupdict()
             if match['list']:
                 url = str(input_url)
-                whole_series(url, current_directory)
+                whole_series(url, current_directory, logger)
             else:
                 pass