+++ comic_dl/sites/gomanga.py
@@ -10,8 +10,9 @@ import sys
 from  more_itertools import unique_everseen
 from bs4 import BeautifulSoup
 from downloader.cookies_required import main as FileDownloader
+import logging
 
-def single_chapter(url,current_directory):
+def single_chapter(url,current_directory, logger):
     
     if not url:
         print("Couldn't get the URL. Please report it on Github Repository.")
@@ -34,6 +35,7 @@ def single_chapter(url,current_directory):
     try:
         chapter_number = int(str(re.search('0\/(.*?)/', url).group(1)).strip().replace('0','').replace('/','')) # Getting the chapter count from the URL itself for naming the folder/dicrectories in integer.
     except Exception as e:
+        logging.debug("Error in Chapter Number : %s" % e)
         chapter_number = 0 # Name the chapter 0 if nothing INTEGER type comes up
 
 
@@ -49,6 +51,7 @@ def single_chapter(url,current_directory):
   
 
     ddl_list = list(unique_everseen(ddl_image_list))
+    logging.debug("Image Links : %s" % ddl_list)
 
 
     print('\n')
@@ -60,14 +63,15 @@ def single_chapter(url,current_directory):
         if not os.path.exists(File_Directory):
                         os.makedirs(File_Directory)
         ddl_image = "http://gomanga.co/reader/content/comics"+str(i).replace('"','').replace('\\','')
+        logging.debug("Image Link : %s" % ddl_image)
 
         File_Name_Final = str(re.findall('\/(\d+)\.[jpg]|[png]', i)).replace("[","").replace("]","").replace("'","").replace(",","").strip()+"."+str(re.findall('\d\.(.*?)$', str(i))).replace(",","").replace("[","").replace("]","").replace("'","").strip()
-        FileDownloader(File_Name_Final,Directory_path,tasty_cookies,ddl_image)
+        FileDownloader(File_Name_Final,Directory_path,tasty_cookies,ddl_image, logger)
 
     print('\n')
     print("Completed downloading ",Series_Name)
 
-def whole_series(url,current_directory):
+def whole_series(url,current_directory, logger):
     if not url:
         print("Couldn't get the URL. Please report it on Github Repository.")
     
@@ -92,9 +96,12 @@ def whole_series(url,current_directory):
         x = link.findAll('a')
         for a in x:
             url = a['href']
-            single_chapter(url,current_directory)
+            logging.debug("Final URL : %s" % url)
+            single_chapter(url,current_directory, logger)
 
-def gomanga_Url_Check(input_url,current_directory):
+def gomanga_Url_Check(input_url,current_directory, logger):
+    if logger == "True":
+        logging.basicConfig(format='%(levelname)s: %(message)s', filename="Error Log.log", level=logging.DEBUG)
     
     gomanga_single_regex = re.compile('https?://(?P<host>gomanga.co)/reader/read/(?P<comic_single>[\d\w-]+)/en/(?P<volume>\d+)?/(?P<Chapter>\d+)?()|(/page/(?P<PageNumber>\d+)?)')
     gomanga_whole_regex = re.compile('^https?://(?P<host>gomanga.co)/reader/(?P<series>series)?/(?P<comic>[\d\w-]+)?(\/|.)$')
@@ -106,7 +113,7 @@ def gomanga_Url_Check(input_url,current_directory):
             match = found.groupdict()
             if match['Chapter']:
                 url = str(input_url)
-                single_chapter(url,current_directory)
+                single_chapter(url,current_directory, logger)
             else:
                 pass
                 
@@ -117,7 +124,7 @@ def gomanga_Url_Check(input_url,current_directory):
             match = found.groupdict()
             if match['comic']:
                 url = str(input_url)
-                whole_series(url,current_directory)
+                whole_series(url,current_directory, logger)
             else:
                 pass