+++ comic_dl/sites/kisscomicus.py
@@ -8,22 +8,21 @@ import os
 # from downloader.universal import main as FileDownloader
 from downloader.cookies_required import main as FileDownloader
 import requests
+import logging
 
-def single_chapter(url, directory):
-    # print("I'm single!")
+def single_chapter(url, directory, logger):
+
     sess = requests.session()
     sess = cfscrape.create_scraper(sess)
     s = sess.get(url)
     cookies = sess.cookies
     connection = s.text.encode('utf-8')
-    # print(cookies)
 
     soup = BeautifulSoup(connection, "html.parser")
     Series_Name_Finder = soup.findAll('ul', {'class': 'back-info'})
     for link in Series_Name_Finder:
         x = link.findAll('a')
         for a in x:
-            # print(a['href'])
             raw_name = a['href']
             Series_Name = str(raw_name.split("/")[2]).replace(".html","").replace("-"," ").title().strip()
             # print("Series Name : %s" % Series_Name)
@@ -44,6 +43,8 @@ def single_chapter(url, directory):
 
     # soup = BeautifulSoup(connection, "html.parser")
     linkFinder = soup.findAll('ul', {'class': 'list-image'})
+    logging.debug("Image Links : %s" % linkFinder)
+
     # print("Link Finder :s %s" % linkFinder)
     for link in linkFinder:
         x = link.findAll('img')
@@ -51,14 +52,15 @@ def single_chapter(url, directory):
             if not os.path.exists(File_Directory):
                 os.makedirs(File_Directory)
             ddlLink = a['src']
+            logging.debug("Final URL : %s" % ddlLink)
             fileName = str(ddlLink).split("/")[-1].strip()
             # print("Link : %s\nFile Name : %s" % (ddlLink, fileName))
-            FileDownloader(File_Name_Final=fileName, Directory_path=File_Directory, tasty_cookies=cookies, ddl_image=ddlLink, )
+            FileDownloader(File_Name_Final=fileName, Directory_path=File_Directory, tasty_cookies=cookies, ddl_image=ddlLink, logger=logger)
 
     print('\n')
     print("Completed downloading %s" % Series_Name)
 
-def whole_series(url, directory):
+def whole_series(url, directory, logger):
 
     scraper = cfscrape.create_scraper()
     connection = scraper.get(url).content
@@ -71,12 +73,15 @@ def whole_series(url, directory):
         for a in x:
             # print(a['href'])
             url = "http://kisscomic.us" + a['href']
-            single_chapter(url, directory)
+            logging.debug("Chapter URL : %s" % url)
+            single_chapter(url, directory, logger)
     print("Finished Downloading")
 
 
 
-def kissmcomicus_Url_Check(input_url, current_directory):
+def kissmcomicus_Url_Check(input_url, current_directory, logger):
+    if logger == "True":
+        logging.basicConfig(format='%(levelname)s: %(message)s', filename="Error Log.log", level=logging.DEBUG)
     kissmcomicus_single_regex = re.compile('https?://(?P<host>[^/]+)/chapters/(?P<comic>[\d\w-]+)(?:/Issue-)?')
     kissmcomicus_whole_regex = re.compile('https?://(?P<host>[^/]+)/comics/(?P<comic_name>[\d\w-]+)?')
 
@@ -87,7 +92,7 @@ def kissmcomicus_Url_Check(input_url, current_directory):
             match = found.groupdict()
             if match['comic']:
                 url = str(input_url)
-                single_chapter(url, current_directory)
+                single_chapter(url, current_directory, logger)
             else:
                 pass
 
@@ -96,6 +101,6 @@ def kissmcomicus_Url_Check(input_url, current_directory):
             match = found.groupdict()
             if match['comic_name']:
                 url = str(input_url)
-                whole_series(url, current_directory)
+                whole_series(url, current_directory, logger)
             else:
                 pass